optimize guest homepage landing image
